+++ src/imio/helpers/content.py
@@ -682,10 +682,10 @@ def object_ids(context, class_names):
     return res
 
 
-def get_user_fullname(username):
+def get_user_fullname(userid, none_if_no_user=False):
     """Get fullname without using getMemberInfo that is slow slow slow...
        We get it only from mutable_properties or authentic.
-       If no fullname, return username instead nothing."""
+       If no fullname, return userid instead nothing."""
     acl_users = api.portal.get_tool('acl_users')
     storages = [acl_users.mutable_properties._storage, ]
     # if authentic is available check it first
@@ -693,14 +693,23 @@ def get_user_fullname(username):
         storages.insert(0, acl_users.authentic._useridentities_by_userid)
 
     for storage in storages:
-        data = storage.get(username, None)
+        data = storage.get(userid, None)
         if data is not None:
             fullname = ""
             # mutable_properties
             if hasattr(data, 'get'):
-                fullname = data.get('fullname')
+                if not data.get('isGroup'):
+                    fullname = data.get('fullname')
+                elif none_if_no_user:
+                    return None
+                else:
+                    return userid
             # authentic
             else:
                 fullname = data._identities['authentic-agents'].data['fullname']
-            return fullname or username
-    return username
+            break
+    else:  # we didn't find this userid
+        if none_if_no_user:
+            return None
+        return userid
+    return fullname or userid